+++ runtime/onert/api/src/nnfw_api_internal.cc
@@ -49,45 +49,6 @@ bool null_terminating(const char *str, uint32_t length)
   return false;
 }
 
-onert::ir::Layout convertLayout(NNFW_LAYOUT layout)
-{
-  if (layout == NNFW_LAYOUT_CHANNELS_LAST)
-  {
-    return onert::ir::Layout::NHWC;
-  }
-  else if (layout == NNFW_LAYOUT_CHANNELS_FIRST)
-  {
-    return onert::ir::Layout::NCHW;
-  }
-  return onert::ir::Layout::UNKNOWN;
-}
-
-NNFW_STATUS getTensorIndexImpl(const onert::ir::Graph &graph, const char *tensorname,
-                               uint32_t *index, bool is_input)
-{
-  if (!tensorname || !index)
-    return NNFW_STATUS_UNEXPECTED_NULL;
-
-  if (!null_terminating(tensorname, MAX_TENSOR_NAME_LENGTH))
-  {
-    std::cerr << "nnpackage path is too long" << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-
-  auto ind_found = is_input ? graph.getInputIndex(tensorname) : graph.getOutputIndex(tensorname);
-
-  if (ind_found.undefined())
-  {
-    // Not found
-    return NNFW_STATUS_ERROR;
-  }
-  else
-  {
-    *index = ind_found.value();
-    return NNFW_STATUS_NO_ERROR;
-  }
-}
-
 } // namespace
 
 nnfw_session::nnfw_session()
@@ -192,310 +153,6 @@ NNFW_STATUS nnfw_session::await()
   return NNFW_STATUS_NO_ERROR;
 }
 
-NNFW_STATUS nnfw_session::set_input(uint32_t index, NNFW_TYPE /*type*/, const void *buffer,
-                                    size_t length)
-{
-  if (!isStatePreparedOrFinishedRun())
-  {
-    std::cerr << "Error during nnfw_session::set_input : invalid state" << std::endl;
-    return NNFW_STATUS_INVALID_STATE;
-  }
-
-  if (!buffer && length != 0)
-  {
-    std::cerr
-      << "Error during nnfw_session::set_input : given buffer is NULL but the length is not 0"
-      << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-
-  try
-  {
-    _execution->setInput(onert::ir::IOIndex(index), buffer, length);
-  }
-  catch (const std::exception &e)
-  {
-    std::cerr << "Error during nnfw_session::set_input : " << e.what() << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-  return NNFW_STATUS_NO_ERROR;
-}
-
-NNFW_STATUS nnfw_session::set_output(uint32_t index, NNFW_TYPE /*type*/, void *buffer,
-                                     size_t length)
-{
-  if (!isStatePreparedOrFinishedRun())
-  {
-    std::cerr << "Error during nnfw_session::set_output : invalid state" << std::endl;
-    return NNFW_STATUS_INVALID_STATE;
-  }
-
-  if (!buffer && length != 0)
-  {
-    std::cerr
-      << "Error during nnfw_session::set_output : given buffer is NULL but the length is not 0"
-      << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-
-  try
-  {
-    _execution->setOutput(onert::ir::IOIndex(index), buffer, length);
-  }
-  catch (const std::exception &e)
-  {
-    std::cerr << "Error during nnfw_session::set_output : " << e.what() << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-  return NNFW_STATUS_NO_ERROR;
-}
-
-NNFW_STATUS nnfw_session::input_size(uint32_t *number)
-{
-  if (isStateInitialized()) // Model is not loaded
-    return NNFW_STATUS_INVALID_STATE;
-
-  try
-  {
-    if (number == nullptr)
-    {
-      std::cerr << "Error during nnfw_session::input_size, number is null pointer." << std::endl;
-      return NNFW_STATUS_UNEXPECTED_NULL;
-    }
-    *number = primary_subgraph()->getInputs().size();
-  }
-  catch (const std::exception &e)
-  {
-    std::cerr << "Error during nnfw_session::input_size : " << e.what() << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-  return NNFW_STATUS_NO_ERROR;
-}
-
-NNFW_STATUS nnfw_session::output_size(uint32_t *number)
-{
-  if (isStateInitialized()) // Model is not loaded
-    return NNFW_STATUS_INVALID_STATE;
-
-  try
-  {
-    if (number == nullptr)
-    {
-      std::cerr << "Error during nnfw_session::output_size, number is null pointer." << std::endl;
-      return NNFW_STATUS_UNEXPECTED_NULL;
-    }
-    *number = primary_subgraph()->getOutputs().size();
-  }
-  catch (const std::exception &e)
-  {
-    std::cerr << "Error during nnfw_session::output_size" << e.what() << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-  return NNFW_STATUS_NO_ERROR;
-}
-
-NNFW_STATUS nnfw_session::set_input_layout(uint32_t index, NNFW_LAYOUT layout)
-{
-  try
-  {
-    if (layout != NNFW_LAYOUT_NONE && layout != NNFW_LAYOUT_CHANNELS_FIRST &&
-        layout != NNFW_LAYOUT_CHANNELS_LAST)
-    {
-      std::cerr << "Error during nnfw_session::set_input_layout, not supported layout" << std::endl;
-      return NNFW_STATUS_ERROR;
-    }
-    _execution->setInputLayout(onert::ir::IOIndex(index), convertLayout(layout));
-  }
-  catch (const std::exception &e)
-  {
-    std::cerr << "Error during nnfw_session::set_input_layout : " << e.what() << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-  return NNFW_STATUS_NO_ERROR;
-}
-
-NNFW_STATUS nnfw_session::set_output_layout(uint32_t index, NNFW_LAYOUT layout)
-{
-  try
-  {
-    if (layout != NNFW_LAYOUT_NONE && layout != NNFW_LAYOUT_CHANNELS_FIRST &&
-        layout != NNFW_LAYOUT_CHANNELS_LAST)
-    {
-      std::cerr << "Error during nnfw_session::set_output_layout, not supported layout"
-                << std::endl;
-      return NNFW_STATUS_ERROR;
-    }
-    _execution->setOutputLayout(onert::ir::IOIndex(index), convertLayout(layout));
-  }
-  catch (const std::exception &e)
-  {
-    std::cerr << "Error during nnfw_session::set_output_layout : " << e.what() << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-  return NNFW_STATUS_NO_ERROR;
-}
-
-static NNFW_TYPE datatype_to_nnfw_dtype(onert::ir::DataType dt)
-{
-  using onert::ir::DataType;
-  switch (dt)
-  {
-    case DataType::FLOAT32:
-      return NNFW_TYPE_TENSOR_FLOAT32;
-    case DataType::INT32:
-      return NNFW_TYPE_TENSOR_INT32;
-    case DataType::QUANT_UINT8_ASYMM:
-      return NNFW_TYPE_TENSOR_QUANT8_ASYMM;
-    case DataType::BOOL8:
-      return NNFW_TYPE_TENSOR_BOOL;
-    case DataType::UINT8:
-      return NNFW_TYPE_TENSOR_UINT8;
-    case DataType::INT64:
-      return NNFW_TYPE_TENSOR_INT64;
-    case DataType::QUANT_INT8_ASYMM:
-      return NNFW_TYPE_TENSOR_QUANT8_ASYMM_SIGNED;
-    case DataType::UINT32:
-    case DataType::QUANT_INT8_SYMM:
-    default:
-      throw std::runtime_error("Error: Model has type that runtime API does not support.");
-  }
-}
-
-NNFW_STATUS nnfw_session::apply_tensorinfo(uint32_t index, nnfw_tensorinfo ti)
-{
-  // sanity check
-  {
-    if (isStateInitialized())
-    {
-      std::cerr << "Error during set_input_tensorinfo : should be run after load_model"
-                << std::endl;
-      return NNFW_STATUS_INVALID_STATE;
-    }
-
-    if (ti.rank <= 0 || ti.rank > NNFW_MAX_RANK)
-    {
-      std::cerr << "unsupported rank: " << ti.rank << std::endl;
-      return NNFW_STATUS_ERROR;
-    }
-
-    for (int32_t i = 0; i < ti.rank; ++i)
-    {
-      if (ti.dims[i] <= 0)
-      {
-        std::cerr << "dim must be positive integer but was " << ti.dims[i] << std::endl;
-        return NNFW_STATUS_ERROR;
-      }
-    }
-  }
-
-  onert::ir::Shape new_shape(ti.rank);
-  for (int32_t i = 0; i < ti.rank; i++)
-    new_shape.dim(i) = ti.dims[i];
-
-  if (!isStatePreparedOrFinishedRun())
-  {
-    // In this case, if we apply input shape in primary_subgraph, it will propagate after
-    // compilation and excution
-    auto primary_subgraph = _subgraphs->primary();
-    auto ind = primary_subgraph->getInputs().at(index);
-    auto &input = primary_subgraph->operands().at(ind);
-
-    // overwrite input shape with the shape from ti
-    input.info().shape(new_shape);
-  }
-  else // when called after nnfw_session::prepare()
-  {
-    _execution->changeInputShape(onert::ir::IOIndex(index), new_shape);
-  }
-
-  return NNFW_STATUS_NO_ERROR;
-}
-
-NNFW_STATUS nnfw_session::set_input_tensorinfo(uint32_t index, const nnfw_tensorinfo *ti)
-{
-  nnfw_tensorinfo ti_copy = *ti;
-  return apply_tensorinfo(index, ti_copy);
-}
-
-NNFW_STATUS nnfw_session::input_tensorinfo(uint32_t index, nnfw_tensorinfo *ti)
-{
-  if (isStateInitialized())
-    return NNFW_STATUS_INVALID_STATE;
-
-  try
-  {
-    if (ti == nullptr)
-    {
-      std::cerr << "Error during nnfw_session::input_tensorinfo, tensorinfo is null pointer."
-                << std::endl;
-      return NNFW_STATUS_UNEXPECTED_NULL;
-    }
-    if (index >= primary_subgraph()->getInputs().size())
-    {
-      std::cerr << "Error during nnfw_session::input_tensorinfo, index is out of range."
-                << std::endl;
-      return NNFW_STATUS_ERROR;
-    }
-    auto opidx = primary_subgraph()->getInputs().at(index);
-    auto shape = primary_subgraph()->operands().at(opidx).shape();
-    if (isStatePreparedOrFinishedRun())
-      shape = _execution->getInputShape(onert::ir::IOIndex{index});
-    ti->rank = shape.rank();
-    for (int j = 0; j < ti->rank; ++j)
-    {
-      ti->dims[j] = shape.dim(j);
-    }
-    ti->dtype = datatype_to_nnfw_dtype(primary_subgraph()->operands().at(opidx).typeInfo().type());
-  }
-  catch (const std::exception &e)
-  {
-    std::cerr << "Error during nnfw_session::input_tensorinfo : " << e.what() << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-  return NNFW_STATUS_NO_ERROR;
-}
-
-NNFW_STATUS nnfw_session::output_tensorinfo(uint32_t index, nnfw_tensorinfo *ti)
-{
-  if (isStateInitialized())
-    return NNFW_STATUS_INVALID_STATE;
-
-  if (ti == nullptr)
-  {
-    std::cerr << "Error during nnfw_session::output_tensorinfo, tensorinfo is null pointer."
-              << std::endl;
-    return NNFW_STATUS_UNEXPECTED_NULL;
-  }
-
-  if (index >= primary_subgraph()->getOutputs().size())
-  {
-    std::cerr << "Error during nnfw_session::output_tensorinfo, index is out of range."
-              << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-
-  try
-  {
-    auto opidx = primary_subgraph()->getOutputs().at(index);
-    auto shape = primary_subgraph()->operands().at(opidx).shape();
-    // If it is called after `nnfw_run` then get the shape from Execution, not from the graph
-    if (isStateFinishedRun())
-      shape = _execution->getOutputShape(onert::ir::IOIndex{index});
-    ti->rank = shape.rank();
-    for (int j = 0; j < ti->rank; ++j)
-    {
-      ti->dims[j] = shape.dim(j);
-    }
-    ti->dtype = datatype_to_nnfw_dtype(primary_subgraph()->operands().at(opidx).typeInfo().type());
-  }
-  catch (const std::exception &e)
-  {
-    std::cerr << "Error during nnfw_session::output_tensorinfo : " << e.what() << std::endl;
-    return NNFW_STATUS_ERROR;
-  }
-
-  return NNFW_STATUS_NO_ERROR;
-}
 NNFW_STATUS nnfw_session::register_custom_operation(const std::string &id,
                                                     nnfw_custom_eval eval_func)
 {
@@ -771,14 +428,4 @@ bool nnfw_session::isStateFinishedRun()
 bool nnfw_session::isStatePreparedOrFinishedRun()
 {
   return isStatePrepared() || isStateFinishedRun();
-}
-
-NNFW_STATUS nnfw_session::input_tensorindex(const char *tensorname, uint32_t *index)
-{
-  return getTensorIndexImpl(*primary_subgraph(), tensorname, index, true);
-}
-
-NNFW_STATUS nnfw_session::output_tensorindex(const char *tensorname, uint32_t *index)
-{
-  return getTensorIndexImpl(*primary_subgraph(), tensorname, index, false);
 }
